safety-tool-parser: rename property_attr module
