+++ demo/rustc_driver/safety-tool-parser/src/lib.rs
@@ -2,4 +2,4 @@ pub use proc_macro2;
 pub use quote;
 pub use syn;
 
-pub mod precond;
+pub mod property_attr;
